feat(courses): add module page and fix lesson redirect
