+++ apps/mobile_app/lib/features/catalog/presentation/widgets/course_header.dart
@@ -1,20 +1,19 @@
 import 'package:flutter/material.dart';
 
-typedef Void = void Function();
-
 class CourseHeader extends StatelessWidget {
   const CourseHeader({
-    required this.title,
-    super.key,
+    required this.title, super.key,
     this.progress = 0,
     this.onContinue,
     this.onBack,
+    this.onTitleTap,
   });
 
   final String title;
   final double progress;
-  final Void? onContinue;
-  final Void? onBack;
+  final VoidCallback? onContinue;
+  final VoidCallback? onBack;
+  final VoidCallback? onTitleTap;
 
   @override
   Widget build(BuildContext context) {
@@ -73,18 +72,24 @@ class CourseHeader extends StatelessWidget {
                   ),
                 ],
               ),
-
               Center(
-                child: Text(
-                  title,
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                    fontWeight: FontWeight.w700,
+                child: GestureDetector(
+                  onTap: onTitleTap,
+                  behavior: HitTestBehavior.opaque,
+                  child: Padding(
+                    padding: const EdgeInsets.symmetric(horizontal: 8, 
+                      vertical: 6),
+                    child: Text(
+                      title,
+                      maxLines: 1,
+                      overflow: TextOverflow.ellipsis,
+                      style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                            fontWeight: FontWeight.w700,
+                          ),
+                    ),
                   ),
                 ),
               ),
-
               Row(
                 mainAxisAlignment: MainAxisAlignment.end,
                 children: [
